requirements-all.txt
# Install everything to try all frameworks
streamlit==1.28.0

# LangChain ecosystem
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.0
langgraph==0.0.20

# Agent frameworks
crewai==0.1.25
crewai-tools==0.0.15
pyautogen==0.2.0

# LLM
ollama==0.1.6
openai==1.3.0

# Tools
duckduckgo-search==4.1.1
wikipedia==1.4.0
arxiv==2.0.0

# Database
sqlalchemy==2.0.23

# Utilities
requests==2.31.0